feat: Add chewie and cached_network_image dependencies
This commit adds the `chewie` and `cached_network_image` dependencies to the `pubspec.yaml` file. These dependencies will be used to improve the video loading and image caching functionality in the app. Additionally, the commit includes some minor code changes in the `MessagesService.dart` and `Chat.dart` files to handle private and group messages respectively.

Please review and merge this commit.